Clarify robot name requirements
Someone brought up the fact that the README suggests that the name function should
be both a query and a command. It was not intended that way, so this makes the
instructions more ambiguous, allowing the name to be defined either during some setup
method, or by making a method that is both a command and a query.