largest-series-product: Comment on tricky cases
Cases involving 0 spans are understandably tricky. Some have argued that
LSP('', 0) should be an error because it takes the max of an empty list.
An example is https://github.com/exercism/xclojure/issues/49.

I agree that taking the max of an empty list is zero.
I disagree that LSP('', 0) takes the max of an empty list.

There is exactly one 0-character string in the empty string (or any
other string), that being the empty string.
The product of that is 1, as the identity element of multiplication.

Since there is reasonable chance for confusion, it is best to add
explanatory comments.