series, largest-series-product: clarify "consecutive"
It was unclear whether "consecutive" meant consecutive in *numerical
value* or simply *positionally adjacent*. Due to this problem's origin
in Project Euler, it was the latter.

To avoid any confusion this commit does the following:
* Replace the word "consecutive" in .yml files with "contiguous"
* Clarifies in the .md files that the series of digits only need be
  positionally adjacent, not consecutive in numerical value.
* Uses examples in .md files that should make clear that numerical value
  is not important.

This should prevent any further confusion of the form expressed in the
two issues

* https://github.com/exercism/xruby/issues/244
* https://github.com/exercism/xruby/issues/262

This is the first subtask of #197.