rna-transcription: remove rna->dna cases from canonical data
There is no biological basis for RNA->DNA transcription.

A couple of people suggested that we could rephrase this and make the
question _Given this RNA, what DNA sequence could produce it?_, but
it seemed simpler to simply remove it.

There is also the question of whether or not having the reverse
transcription helps guide people to a better/more generic solution.
I think that this is one place where we could have the discussion on the
website about hard-coding vs not.

See https://github.com/exercism/x-common/issues/148