Adding Markdown refactoring exercise
It would be great to have some more refactoring exercises, and one that is easy
to write both poorly and well is a Markdown parser. This is just the outline
for that, but it does include a pretty basic set of tests that cover a small
subset of Markdown syntax.