tournament: clarify behavior on invalid lines (ignore)
"Only accept" was a bit vague - this commit changes the mesage to say
that invalid line are ignored, and the tournament scored without them.

This is in line with the behavior of all current tracks. If reviewers
wish a change to this behavior, we can discuss that in a separate PR.

(Note that the go track is an exception, but it clearly states it is an
exception in its test suite visible to students)

Closes #286
Closes https://github.com/exercism/xrust/issues/122