tournament: Add JSON file
For the most part, the existing implementations all used the same tests:

* https://github.com/exercism/xcsharp/blob/master/exercises/tournament/TournamentTest.cs
* https://github.com/exercism/xfsharp/blob/master/exercises/tournament/TournamentTest.fs
* https://github.com/exercism/xgo/blob/master/exercises/tournament/tournament_test.go
* https://github.com/exercism/xrust/tree/master/exercises/tournament/tests
* https://github.com/exercism/xswift/blob/master/exercises/tournament/TournamentTest.swift

xgo is the clear outlier: Whereas all other tracks ignore invalid lines,
xgo declares that inputs with invalid lines cause an error (but empty
lines and comments are okay). Please refer to discussion in #286 where
we decide how to make this consistent across tracks.

For now, these tests follow the behavior of the other tracks.

xgo also had a test for ties which no other track had, so that's been
added. This tests that ties are broken alphabetically.

Behaviors not tested:
* Empty input.
* Team playing itself.

Closes https://github.com/exercism/todo/issues/155