rectangles.json: Explain purpose of large input test
As far as I can tell, it was added to rule out inefficient solutions
since it doesn't test any edge cases not already covered in previous
tests. (And even if it did, I would want to just add previous tests that
test those edge cases).

It's useful to explain the purpose of each test so tracks can decide
whether to test for performance or not. Discussion at
https://github.com/exercism/xrust/pull/165 expands on the merits.

Currently the fact that this is a test primarily for performance is not
encoded in a machine-readable way, so tracks that auto-generate tests
will not be able to mark that it's a performance test automatically. We
can revisit that if there is a large demand for it.