Explicitly describe robot name format (#283)
The definition of this problem led to some confusion in the scala test case. Some languages expect differing formats for the robots names. This is an attempt to make it explicit and clear to the end-users and the problem creators to avoid this ambiguity.

An example of this issue can be found in this PR: https://github.com/exercism/xscala/pull/123

If this PR is accepted, I will be more than happy to check the languages that have the "Robot Name" problem to make sure each language's tests are consistent.