tournament: Remove tiebreak in favor of more wins
Note that this sentence was added in #254 and previously was not
present. No track has a test case that tests this condition (equal
points, unequal number of wins).

Because of a combination of various factors:
* Teams play each other no more than once
* Four teams per group
* Wins award three points and draws one

It's not possible to have a tie with teams with unequal numbers of wins.
At the minimum:
* One team (A) would get one win and zero draws.
* One team (B) would get zero wins and three draws.

That's not possible, since at least one draw from team B would be
against team A, giving team A an extra point and breaking the tie.

We *can* solve this by breaking any one of the above three conditions.
If we would like to, we should make that explicit decision and edit the
JSON test cases to match. However, on the assumption that we want to
keep things as they are (and were before #254), this commit removes the
sentence.

This was the suggested course of action in #287.