anagram: rm all "string argument" cases
the ant -> stand, tan, at case is exactly a duplicate of a case in the
"cases" section. The ant -> tan case is not a duplicate, but it does not
seem to provide any extra value as the description of the test "accepts
single string argument" is covered by, for example, galea -> eagle.