hello-world: Be more function-oriented, not program-oriented
As https://github.com/exercism/exercism.io/issues/2746 rightfully
explains, the wording of "program" implies that the student is expected
to write a full program with a main function (or equivalent entry point
in language of choice).

Further, the wording of "greets" slightly implies that the function is
expected to put the result on STDOUT.

To remedy this, "program" is replaced with "function" and we specify
that the function should return the greeting.

Note that as #321 states, this is just one of many problems that has
"write a program" in its description. However, the "greets" -> "returns"
wording is an issue not covered by #321, so we deal with it here.

Closes https://github.com/exercism/exercism.io/issues/2746