transpose: use arrays for inputs/outputs
For the same motivation as [in grep][1], it seems advantageous to have
everything use arrays, in case there are languages that wish to
automatically parse the JSON file. Such languages may have a difficult
time if the type can sometimes be strings but other times be arrays;
always using arrays makes it consistent.

[1]: https://github.com/exercism/x-common/pull/327#discussion_r73994597