pov: Add canonical data
Existing implementations:

* https://github.com/exercism/xclojure/blob/master/exercises/pov/test/pov_test.clj
* https://github.com/exercism/xfsharp/blob/master/exercises/pov/PovTest.fs
* https://github.com/exercism/xgo/blob/master/exercises/pov/pov_test.go
* https://github.com/exercism/xhaskell/blob/master/exercises/pov/test/Tests.hs
* https://github.com/exercism/xlua/blob/master/exercises/pov/pov_spec.lua

The tests are a combination of the existing tests.

The trees used:

* Singleton (present in all)
* Simple tree: X, parent, sibling (present in Clojure, Go, Lua)
* Large flat tree: X, parent, three siblings (present in all)
* Deeply nested tree: "singly-linked list" as a tree (present in all)
* Target has children: X, parent, two children (present in all*)
  (in Clojure, Go, Lua, this case is embedded in a more complex tree)
* Target has cousins (present in all)

All tracks tested looking for the node named "x" in these trees, then
looking for a nonexistent node.

The Lua track had a gentler progression of complexity in the path tests,
so that was used. This is necessary simply because even after the reroot
is done, there is still work to be done to search for the target node.

(The Go track chose not to test paths)

Closes https://github.com/exercism/todo/issues/128