Grains: Add canonical-data.json file
Current implementations of this exercise are all nearly identical, so
creating a canonical test set was just a matter of documenting the
existing tests.

Some languages do not include the error-handling tests, but I think they
are valuable to include as the boundaries they test are part of the
problem domain.