[Bowling] Add edge case with fill balls in a last-frame strike (#418) (#423)
* Add edge case with fill balls in a last-frame strike

The previous test does not quite capture the behavior.

If the first fill fill is a non-strikes, then the total fill-ball score
must be less than 10.

But, if the first fill ball is a strike, then the total fill-ball score
must be less than 20. Because that first fill-ball strike resets the
pins.

I had what I thought was a working implementation but it totally missed
the 2nd case. Adding a test to cover that case.

* Add edge case test for final-strike fill balls

This test checks that the fill balls validate correctly if the 2nd ball
is a strike.