Triangle: Update tests to check properties, not types
As discussed here: https://github.com/exercism/x-common/issues/379

Converts the tests so that they are checking the property of a triangle
and not its type (or type stand-in like symbol/atom/etc.).

I've removed some duplicative tests, but I believe I've still covered
all the edge cases that were part of the old test suite.