food-chain: resolve ambiguity in multiple verses test
The current test for multiple verses checks for two consecutive
verses (from 1 to 2) which leaves room for interpretation.

It is unclear whether the solution should simply concatenate two (or
more, when variadic functions are available) possibly non-consecutive
verses or return a range of consecutive verses from start to end. As
observed in the Go track, this ambiguity results in differing
solutions which makes them harder to compare.

This commit changes the multiple verses test to cover two non-
consecutive verses (1 and 3). With the change, only a solution
returning a range of consecutive verses is accepted.

See exercism/xgo#412