Update Luhn tests and description
I'm addressing a few different things here:

1. The tests exposed functionality that is not needed

The `addends` and `check_digit` methods aren't needed by clients of this
code, and I didn't see that students got any benefit from explicitly
implementing them. The presence of these tests pushed students towards a
particular implementation, which we [try to
avoid](https://github.com/exercism/x-common/blob/master/CONTRIBUTING.md#improving-consistency-by-extracting-shared-test-data)

> The test cases should not require people to follow one very specific
way to solve the problem, e.g. the tests should avoid checking for
helper functions that might not be present in other implementations.

To fix this I've removed tests of `addends` and `check_digit`.

2. The tests were duplicative.

There's not a way (that I could see, at least) of implementing Luhn in
tiny steps. It's either working or its not. Once you have a couple valid
tests and a couple invalid tests your are pretty much covered. (note: I
guess you could test the cases where Luhn can be wrong, but that seemed
outside the scope of this exercise.)

Again, from the guide

> All tests should be essential to the problem. Ensure people can get
the first test passing easily. Then each following test should ideally
add a little functionality. People should be able to solve the problem a
little at a time. Avoid test cases that don't introduce any new aspect
and would already pass anyway.

To fix this I've removed most duplicative tests that would pass as soon
as the student got a working Luhn implementation.

3. The examples and the description didn't match

Luhn can be hard to grasp. At least it can be for me. What helps me is
to have explicit, real-world examples. To help with this I've updated
the description.md file to have examples of credit cards and Canadian
SINs, two places where Luhn is used in the real world.

To re-enforce the descriptions, I've used the same examples in the
tests. So students can see step-by-step breakdowns of how to get their
tests passing.

Note:

I'm still on the fence about the `generate_check_digit` method. It seems
to be to be outside the scope of this exercise, but I also think
implementing it is interesting and useful. I could go either way on
keeping it or removing it.