exercise: book-store, canonical data (#470)
* Create harry-potter

* Create description.md

* Delete description.md

* Delete harry-potter

* Create description.md

* canonical data for book-store

initial version

* Updating canonical data

* updated information

Another attempt at re-writing descriptions.  Also converted baskets to arrays.

* Update canonical data

Trying to follow @peterseng suggestions.

* Updated description

Descriptions was too specific

* Fixed spacing concern

* Rewrote decriptions

I was not understanding the the descriptions should relate the solution
that achieves the expected result.  I am not sure having the
descriptions this wordy is desirable.

* Removed all tab characters

Removed all tabs and replaced with appropriate number of spaces.

* Obsolete

* Reducing verbage

Cut down the wordiness of each description.  Introduced "targetgrouping"
field.  This field contains the grouping that yields the most discount
to the customer.

* Second test had wrong targetgrouping

* Two updates

"targetgrouping" should be an array of arrays.  Also modified two
descriptions to make this exercise clearer to the reader.

* Fixed empty array of empty array