variable-length-quantity: Limit to 32-bit int, not 28-bit int (#487)
* variable-length-quantity: s/delta-time/integer/

delta-time seems to only be meaningful in the context of MIDIs. We are
dealing with just arbitrary integers, with no mention of domain.

* variable-length-quantity: Limit to 32-bit int, not 28-bit int

The limit of a 28-bit int (0x0fffffff) is specific to MIDIs, whereas
variable-length quantities can deal with arbitrary sizes.

So why limit to 32-bits? Because it is a common size for languages, and
some language tracks already have tests that assume that the limit is
0xffffffff, so we might as well match them.

Closes #485
Closes https://github.com/exercism/xrust/issues/224