variable-length-quantity: Don't test 32-bit overflow
Since the description has guaranteed to the student that we only need to
restrict ourselves to 32-bit integers, you could argue it's
contradictory to then turn around and test a larger number (despite the
fact that the test expects an error). Instead, leave it undefined. This
way, students have the freedom to make their code handle larger numbers
if desired.