circular-buffer: Add should_succeed to read and write
This makes it clear that there are only four operations, instead of
implying that `readShouldFail` and `writeShouldFail` are potentially
separate operations.

I convinced myself to do this since I argued against a
`failure_detected` as a separate operation, but then realized that
`readShouldFail` and `writeShouldFail` suffer from the same problems.