secret-handshake: Use 3 and 19, not 9 and "11001"
We'd like to use 3 and 19 to match up with what's in the test cases
already used by most languages. We also do not want to do binary to
integer conversions, by discussion in #159.

This covers most tasks in #335, though I would probably still leave #335
open because explicit discussion of numbers larger than 11111 has not
been added.