Luhn: Update test of non-digit
While implementing these tests for Rust I found that the final test was
not well written (my fault!)

https://github.com/exercism/xrust/pull/247

I added the alphabetical character to an already-invalid string. So the
test doesn't show anything useful.

Instead the test should show that a known-valid string is made invalid
by the inclusion of a non-digit