scrabble-score: add alphabet completeness test (#516)
After the discussion in #509 , here is proposed addition to the
canonical data.

If we really wish to stick to "proper, scrabble-approved word", I could
make multiple cases for
"the"
"quick"
"brown"
"fox"
"jumps"
"over"
"lazy"
"dog"

But in my personal opinion that obscures the intended purpose. The
complete alphabet is the shortest way to express the complete alphabet,
and thereby makes the intent the clearest. Anything else just muddles
the waters (IMHO).

Closes #509