hello-world: remove reference to "name not given"
This implements the second option "change the hello-world README to be
so vague..." of #520.

As stated in #520, the point of hello-world is to give a very easy
exercise. This helps give an intro to how to fetch, how to submit, and
how to run the test suite for a given language. The README included with
hello-world helps out with that.

The clause "If a name is not given" is very difficult for languages that
do not support overloading functions and/or default function parameters.

This is against the vision of a very easy hello-world. The ideal is a
function that returns "Hello, World!" unconditionally.

By not mentioning the "name not given" in the README, every track that
currently has a hello-world exercise can change it or not, as they
please.

This is a very delicate balance to strike. If we make it too vague, that
is also undesirable. See #322. I cannot be sure that I have gotten it
right in this commit.