change: Add case for target > coins, but still can't make
The current error cases ask us to ensure that two types of targets are
invalid:
* targets smaller than the smallest coin
* negative targets

These are both reasonable.

I would react with chagrin if this led student solutions to assume that
these are the only two cases in which one could fail to meet the target
i.e. they assume that any target > the smallest coin is valid.

This commit proposes adding a case that disproves that - the target is
larger than even the largest coin, but you will not be able to get a
combination that reaches the target.