bowling: be more explicit about when errors should happen
The JSON file currently says "When to error is also left up to your
implementation", but we should instead be explicit.

Let's say I `roll 5`, then `roll 6`, then `roll 0` 18 times.
Okay, that's 20 rolls, that's a full game, so I can take the score now,
right?
Now I try to take the score and find that I get an error. Argh!

In the absolute best case, the error result from score would be so kind
to tell me that my `roll 6` way back there was invalid, but in the worst
case it might not, and I would be left to my own devices to figure out
what happened, since there was no feedback on any of the individual
rolls.

This commit proposes that invalid rolls be made explicit right when the
roll is made, not at some unspecified time in the future (when
attempting to take the score of the game). This is done via a
`roll_should_fail` key that is mutually exclusive with the `expected`
key.

Previous discussions I know of (since I was involved):
* https://github.com/exercism/xrust/pull/213#issuecomment-250988227
* https://github.com/exercism/xhaskell/issues/440