book-store: Explain importance of 4+4 vs 5+3 case
You cannot greedily add books to a single group in this case. You have
to try the possibilities and see that two groups of four is a cheaper
grouping than a group of five and a group of three.

If this is not made clear, people may make assumptions that greedy is
the correct way to go, as in:
https://github.com/exercism/xcsharp/issues/181