sum-of-multiples: Add descriptions to tests
The new JSON Schema for 'canonical-data.json' files
demands that all tests have descriptions, so that
users can know where they failed.

These descriptions may not be the best possible, as
they only explain what are the tests - and not why
they are there - but they are needed for compliance.