triangle: Rewrite comments to fit in 80 columns
Restructure all the comments to fit nicely in 80 columns.
The content is expected to be exactly the same.

No changes made to the descriptions that do not fit the
witdh.  They are supposed to be short, single line texts
to be displayed to the users, but the purpose here is not
to change the contents of the test suite.