beer-song: Make canonical data more logical
We were testing:
- the first verse
- (*) some random verse in the middle
- each of the special case verses
- (*) 4 consecutive verses
- all verses

The two that are marked with an asterisk are not very logical choices in
terms of TDD.

The first asterisk is about the generation of individual verses.

The structure of the algorithm is that verses 99 through 3 all have the
same implementation. Then verses 2, 1, and 0 are all unique.

Instead of testing the first generic verse and some random generic
verse, it makes more sense to test the outer bounds: the first (99) and
the last (3).

The second asterisk is about generating consecutive verses.

When testing multiples, the natural progression isn't "4" then "all".

The best test data is _the smallest number that is multiple_, which is
two. Since that can be implemented with a simple concatenation of the
upper and lower bounds (verse of N + verse of M), it's useful to add
an extra case that pushes towards a loop.

Again, we want the smallest number that suggests a loop, which is three
consecutive verses, not four.

Then, finally, we have the full integration test, which is all the verses.