pov 1.1.0: replace unnecessary uses of "cousins" tree with smaller tree
We would like to reduce the tree so that:

1. It's clearer what is being tested
2. The duplication in the test file is kept to fewer lines

This helps us be rid of the top-level `"trees"` key as desired in #699.

The usages of "cousin" that were replaced:

**Three error cases**:

1. "Errors if target does not exist in a large tree"
2. "Errors if destination does not exist"
3. "Errors if source does not exist"

In all three cases, a "large" tree is all that is necessary. It doesn't
necessarily have to be the entire "cousin" tree.

Besides, it's not very interesting to see whether a student can *fail*
to find a nonexistent node in a large tree, because the student can just
as easily achieve it by not doing any searching at all. It's much more
intersting to see whether the student can find an existent node in a
large tree. So these error cases really don't need a very large tree.

A "large" tree is all that is needed, but it doesn't necessarily have to
so big so as to be the entire "cousin" tree.

**One case of "Can find path from nodes other than x"**:

Nothing in this case requires the use of the specific "cousins" tree.

Since the test contents have now changed, this is now version 1.1.0.