grep: Move files section to comments
Move the 'files' object representing the three files that must be
present when running the tests to the comments, because the new
JSON schema doesn't allow shared data among test cases.

This is the only remaining exercise having a shared data structure,
and there are no plans at the moment to add support for this
feature, so are we moving the data to preserve the information and
allow compliance with the new standard.