Add case where word starts with a unique letter
I used this exercise in an entry-level Python course and a participant
came up with the following attempt:

    def is_isogram(s):
      if s == '':
        return True
      t = s.lower()
      for c in t:
        if t.count(c) > 1:
          return False
        else:
          return True

Note how the loop is unconditionally aborted after the first iteration.
However, all the test cases of the existing test suite run green with
this wrong solution.

The added test case fixes this, i.e. the above fails and the example
solution still succeeds.