number: Rewrite description and add tests
Removes the part about how to test, as we have other docs for that.
Now it also describes the telephone number system in a bit more detail.

Changes some test that used an invalid number but assumed it was valid.

Adds two tests that check the range of 2-9 for the first digit of the area
and the exchange code.

Adds a test with a valid telephone number with country code and punctuation.