word-search: add canonical data (#755)
A single grid is used for all tests. The likelihood that a practitioner will hard-code the actual search logic in any way is low.

The grid used for all tests matches the updated grid from the exercise description.md.

Test cases for each of the 8 possible word directions are cumulative, which allows better matching of the problem statement ("given a puzzle and a list of words") and a more explicit prompt to "build up" the search capability piece-by-piece.