rail-fence-cipher 1.0.1: Remove "test to" from all descriptions
It doesn't seem to add value to state that these are tests! That should
be self-evidence from their presence in this file, no?

I don't have evidence that this a documented principle that MUST be
followed, nor do I have evidence that anyone is being hurt by the
presence of the words "test to", but I propose it to see what people
think.