Prepend blurb to description.
The problem description given in description.md should be complete, self
contained, and not require the addition of the blurb to make sense.

This patch prepends the blurb and a newline to all the descriptions that
did not already begin with the blurb text.

This matches the output of the problem readme as supplied by trackler.