difference-of-squares: add tests for input == 0
Although it's probably hard to come up with a solution that works for all cases
but the 0 one, it should be tested for nevertheless.

When testing units that do arithmetic, 0 as input is a special case that
surprisingly often is forgotten about. And in more complicated code this could
matter.