collatz-conjecture: add test case for a large number (#822)
* collatz-conjecture: add test case

Largest number of steps for number less than one million.  This is loosely based on the Euler Project.  It asks for the number less than 1 million that has the most steps.

* collatz-conjecture: update proposed new case

New case no longer is a spoiler for Euler Project implementation.  Description reworded.  This case is essentially the same as the case above it.  Would it be better to just alter the case above instead?