circular-buffer: Clarify overwrite (#892)
As discussed in #889, the documentation for the overwrite operation
was vague. overwrite replaces the oldest element in the buffer if the
buffer is full and is just a normal write if there is space available.

Overwrite replaces the oldest item in the buffer if the buffer is
empty. Otherwise it is just like write. This commit documents what
overwrite on a buffer with available space looks like and clarifies
what the oldest element is after an overwrite operation succeeds.

Closes #889